+++ components/sensor/collector.go
@@ -6,7 +6,6 @@ import (
 
 	pb "go.viam.com/api/common/v1"
 	"google.golang.org/protobuf/types/known/anypb"
-	"google.golang.org/protobuf/types/known/structpb"
 
 	"go.viam.com/rdk/data"
 )
@@ -24,9 +23,9 @@ func (m method) String() string {
 	return "Unknown"
 }
 
-// NewSensorCollector returns a collector to register a sensor reading method. If one is already registered
+// NewReadingsCollector returns a collector to register a sensor reading method. If one is already registered
 // with the same MethodMetadata it will panic.
-func NewSensorCollector(resource interface{}, params data.CollectorParams) (data.Collector, error) {
+func NewReadingsCollector(resource interface{}, params data.CollectorParams) (data.Collector, error) {
 	sensorResource, err := assertSensor(resource)
 	if err != nil {
 		return nil, err
@@ -42,13 +41,9 @@ func NewSensorCollector(resource interface{}, params data.CollectorParams) (data
 			}
 			return nil, data.FailedToReadErr(params.ComponentName, readings.String(), err)
 		}
-		readings := make(map[string]*structpb.Value)
-		for name, value := range values {
-			val, err := structpb.NewValue(value)
-			if err != nil {
-				return nil, err
-			}
-			readings[name] = val
+		readings, err := data.StructValueMapFromInterfaceMap(values)
+		if err != nil {
+			return nil, err
 		}
 		return pb.GetReadingsResponse{
 			Readings: readings,
