+++ components/sensor/collector_test.go
@@ -10,7 +10,6 @@ import (
 
 	"go.viam.com/rdk/components/sensor"
 	"go.viam.com/rdk/data"
-	du "go.viam.com/rdk/data/testutils"
 	"go.viam.com/rdk/logging"
 	tu "go.viam.com/rdk/testutils"
 	"go.viam.com/rdk/testutils/inject"
@@ -32,7 +31,7 @@ func TestSensorCollector(t *testing.T) {
 	}
 
 	sens := newSensor()
-	col, err := sensor.NewSensorCollector(sens, params)
+	col, err := sensor.NewReadingsCollector(sens, params)
 	test.That(t, err, test.ShouldBeNil)
 
 	defer col.Close()
@@ -40,7 +39,7 @@ func TestSensorCollector(t *testing.T) {
 	mockClock.Add(captureInterval)
 
 	test.That(t, buf.Length(), test.ShouldEqual, 1)
-	test.That(t, buf.Writes[0].GetStruct().AsMap(), test.ShouldResemble, du.GetExpectedReadingsStruct(readingMap).AsMap())
+	test.That(t, buf.Writes[0].GetStruct().AsMap(), test.ShouldResemble, data.GetExpectedReadingsStruct(readingMap).AsMap())
 }
 
 func newSensor() sensor.Sensor {
