+++ components/sensor/sensor.go
@@ -19,7 +19,7 @@ func init() {
 	data.RegisterCollector(data.MethodMetadata{
 		API:        API,
 		MethodName: readings.String(),
-	}, NewSensorCollector)
+	}, NewReadingsCollector)
 }
 
 // SubtypeName is a constant that identifies the component resource API string "Sensor".
